sandbox: Thanks to Francesc Altec suggestion, the loess routines now run without crashing...However, they are not yet fully tested.

+++ Lib/sandbox/pyloess/sandbox/src/loessc.c
@@ -13,7 +13,7 @@ static double *v;
 loess_raw(double *y, double *x, double *weights, double *robust, int *d,
           int*n, double *span, int *degree, int *nonparametric,
           int *drop_square, int *sum_drop_sqr, double *cell, char **surf_stat,
-          double *surface, int *parameter, int *a, double *xi, double *vert,
+          double *surface, long *parameter, long *a, double *xi, double *vert,
           double *vval, double *diagonal, double *trL, double *one_delta,
           double *two_delta, int *setLf)
 {
@@ -116,7 +116,7 @@ loess_dfitse(double *y, double *x, double *x_evaluate, double *weights,
     loess_free();
 }
 
-loess_ifit(int *parameter, int *a, double *xi, double *vert, double *vval,
+loess_ifit(long *parameter, long *a, double *xi, double *vert, double *vval,
            int *m, double *x_evaluate, double *fit)
 {
     loess_grow(parameter, a, xi, vert, vval);
@@ -165,7 +165,7 @@ loess_workspace(int *d, int *n, double *span, int *degree,
         iv[i + 40] = drop_square[i];
 }
 
-loess_prune(int *parameter, int *a, double *xi, double *vert, double *vval)
+loess_prune(long *parameter, long *a, double *xi, double *vert, double *vval)
 {
     int    d, vc, a1, v1, xi1, vv1, nc, nv, nvmax, i, j, k;
 
@@ -198,7 +198,7 @@ loess_prune(int *parameter, int *a, double *xi, double *vert, double *vval)
         vval[i] = v[vv1 + i];
 }
 
-loess_grow(int *parameter, int *a, double *xi, double *vert, double *vval)
+loess_grow(long *parameter, long *a, double *xi, double *vert, double *vval)
 {
     int    d, vc, nc, nv, a1, v1, xi1, vv1, i, j, k;
